Remove resolving from the filter so that version ranges and version pinning properly get propagated.
